requirements.txt
anthropic
certifi==2026.2.25
charset-normalizer==3.4.6
flask==3.1.3
idna==3.11
psutil==6.1.1
python-dotenv==1.2.2
requests==2.32.5
urllib3==2.6.3
